backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
convertdate==2.3.2
python-docx==0.8.11
qrcode==7.3.1
pandas==1.3.3
python-dotenv==0.19.0 